\name{facopyEnrichment}
\alias{facopyEnrichment}

\title{
Gene-set Enrichment Analysis
}
\description{
Performs gene-set enrichment analysis of cancer copy number data, based on a gene table, a \code{facopyInfo} object and the correlations between gene copy number and expression data.
}
\usage{
facopyEnrichment(fad, geneTable, cor, outFolder, 
                 pvalThr = 0.05, corThr = 0.1, plotThr=0.001)
}

\arguments{
  \item{fad}{
\code{facopyInfo} object with a certain study's \code{facopy} data.
}
  \item{geneTable}{
A table with the format of \code{\link{facopy}} function's output, containing at least the columns regarding feature name, p-value and chromosome arm.
}
  \item{cor}{
A list generated by a call to \code{\link{calculateCor}}, containing information on correlation between gene copy number and expression data.
}
  \item{outFolder}{
The folder to which to write the enrichment results.
}
  \item{pvalThr}{
Maximum p-value required to consider a gene to be significantly associated.
}
  \item{corThr}{
Minimum R2 required to consider a gene's copy number and expression to be sufficiently correlated.
}
  \item{plotThr}{
Pathways from Biocarta, KEGG and Reactome with lower p-values than this integer will have their graphs plotted in the output folder. 
}
}
\details{
Only use this function and \code{\link{calculateCor}} if you selected some kind of gene collection as genomic features.
}
\value{
The input gene table is returned with correlation values attached, ordered by p-value and then by R2.
}
\author{
David Mosen-Ansorena
}

\examples{
data(myStudy) # load example study

# then perform an association analysis
# genes = facopy(myStudy, "amp", "stage")

# calculate expression-CN correlations
# eCor = calculateCor(myStudy, "mrna_merged_median_Zscores", "coadread_tcga_pub")

# and run facopy enrichment
# facopyEnrichment(myStudy, genes, eCor, "~/myFolder/stageAmpEnrichment")
}

\keyword{ enrichment }
