\name{facopyInfo-class}
\Rdversion{1.1}
\docType{class}
\alias{facopyInfo-class}
\alias{facopyInfo}
\alias{summary,facopyInfo-method}

\title{Class \code{"facopyInfo"}}
\description{
\code{facopyInfo} is the main unit of information used through the analysis of your data with \code{facopy}.
}
\section{Content}{
 These objects are generated by methods in the package and their basic information can be recovered by the \code{summary} method.
}
\author{
David Mosen-Ansorena
}
\examples{
showClass("facopyInfo")
}
\keyword{classes}
