\name{getFacopyInfo}
\alias{getFacopyInfo}

\title{
List Available facopy Input and Characteristics
}
\description{
Lists of supported items are printed for reference.
}
\usage{
getFacopyInfo()
}
\details{
The following lists of supported items are printed: variable types, alteration combinations, input formats for CN calls, genome builds, genomic features.
}

\author{
David Mosen-Ansorena
}
\examples{
getFacopyInfo()
}

\keyword{ overview }