\name{myCalls}
\alias{myCalls}
\docType{data}
\title{
Example facopy Object with Just Copy Number Calls
}
\description{
Example \code{facopyInfo} object with just the copy number calls, generated using the \code{\link{readCNData}} function.
}
\usage{data(myCalls)}
\format{
  The format is: formal class 'facopyInfo'
}
\examples{
data(myCalls)

summary(myCalls)
}
\keyword{datasets}
