\name{myStudy}
\alias{myStudy}
\docType{data}
\title{
Example Complete facopy Object
}
\description{
Example \code{facopyInfo} object, generated using the \code{\link{readCNData}}, \code{\link{addVariables}} and \code{\link{addFeatures}} functions.
}
\usage{data(myStudy)}
\format{
  The format is: formal class 'facopyInfo'
}
\examples{
data(myStudy)

summary(myStudy)
}
\keyword{datasets}
