\name{myVariables}
\alias{myVariables}
\docType{data}
\title{
Example Phenotypic Annotation
}
\description{
This \code{\link{data.frame}} contains three variables with annotation on phenotypes for 20 samples.
}
\usage{data(myVariables)}
\format{
  A data frame with 20 observations (samples) on the following 3 variables.
  \describe{
    \code{code}: a numeric vector\cr
    \code{age}: a numeric vector\cr
    \code{stage}: a numeric vector
  }
}
\examples{
data(myVariables)

head(myVariables)
}
\keyword{datasets}
