\name{readCNData}
\alias{readCNData}

\title{
Read Copy Number Data
}
\description{
The function reads the copy number output generated by one of the supported copy number analysis tools
}
\usage{
readCNData(folder, method = NULL, sex = c("X"), FUN = NULL, version = NULL, 
           window = NULL, rankThr = NULL, pfbFilename, lengthThr = 10, ...)
}

\arguments{
  \item{folder}{
A character indicating the path where the files with copy number information are located. In order to avoid possible conflicts, try not to place any additional files within the folder.
}
  \item{method}{
Either \code{NULL} or a character.\cr
If a character, it specifies the method used generate the copy number output. Currently supported: \code{"seqcna"}, \code{"cnanorm"}, \code{"patchwork"}, \code{"freec"}, \code{"oncosnp"}, \code{"oncosnp-seq"}, \code{"gap"} and \code{"exomecnv"}.\cr
If \code{NULL}, a generic input will be assumed with the following columns: chromosome, segment start, segment end, segment length and segment type. The segment types should be coded as 1 (deletion), 2 (normal), 3 (amplification), 11 (LOH deletion), 12 (LOH normal) or 13 (LOH amplification).
}
  \item{sex}{
Either \code{NULL} or a character vector. If a character vector, it should contain the names of the sexual chromosomes that should be read.
}
  \item{FUN}{
A pattern replacement function. Some copy number analysis tools add suffixes and file extension characters to the names of their output files. This function allows keeping just the sample names. Leave to \code{NULL} for default replacement based on the selected method. 
}
  \item{version}{
A character with the software version of the copy number analysis tool. The output may vary between versions of the same method. Leave to \code{NULL} if unknown and check your software version if reading the copy number data is not successful.
}
  \item{window}{
An integer. Required for output files from CNAnorm, which do not contain information on the window length used during the copy number analysis process.
}
  \item{rankThr}{
An integer. Threshold on the Rank parameter provided by OncoSNP-SEQ and the newer versions of OncoSNP (>1.3).
}
  \item{pfbFilename}{
For GAP, name of a file with population B allele frequencies (PFB), the same one used during the copy number analysis process.
}
  \item{lengthThr}{
An integer. For GAP, filter those alterations that span less SNPs than this threshold.
}
  \item{\dots}{
If a method is not selected (\code{NULL}), further arguments passed to the \code{read.delim} function, in charge of reading the input files from the selected folder.
}
}

\value{
A \code{facopyInfo} object with copy number data.
}

\author{
David Mosen-Ansorena
}

\examples{
# myCalls = readCNData("~/myFolder/", "seqcna")
# myCalls = readCNData("~/myFolder/", "gap", pfbFilename="~/myPfb.pfb")
# myCalls = readCNData("~/myFolder/", "cnanorm", window=50000)
}

\keyword{ preprocessing }
